Removed unused multithreading functions and updated to uint64_t

+++ client/src/cmdhficlass.h
@@ -43,5 +43,5 @@ uint32_t picopass_elite_rng(void);
 uint32_t picopass_elite_lcg(void);
 uint8_t picopass_elite_nextByte(void);
 void *generate_key_blocks(void *arg);
-void generate_key_block_inverted(const uint8_t *startingKey, uint32_t index, uint8_t *keyBlock);
+void generate_key_block_inverted(const uint8_t *startingKey, uint64_t index, uint8_t *keyBlock);
 #endif
